sangre

+++ AgroRobotView/frmMantSangre.h
@@ -24,6 +24,9 @@ namespace AgroRobotView {
 		{
 			InitializeComponent();
 			this->muestraController = gcnew MuestraController();
+			this->gestorNutricionalController = gcnew GestorNutricionalController();
+			cargarAnimalesDisponibles();
+
 			this->BackColor = System::Drawing::Color::SeaGreen;
 			this->button1->BackColor = System::Drawing::Color::LightGreen;
 			this->button1->ForeColor = System::Drawing::Color::DarkGreen;
@@ -56,7 +59,7 @@ namespace AgroRobotView {
 
 
 
-
+	private: GestorNutricionalController^ gestorNutricionalController;
 	private: System::Windows::Forms::GroupBox^ groupBox1;
 	private: System::Windows::Forms::Button^ button1;
 	private: System::Windows::Forms::TextBox^ textBox1;
@@ -70,6 +73,14 @@ namespace AgroRobotView {
 	private: System::Windows::Forms::DataGridViewTextBoxColumn^ Column5;
 	private: System::Windows::Forms::DataGridViewTextBoxColumn^ Column6;
 	private: System::Windows::Forms::DataGridViewTextBoxColumn^ Column7;
+	private: System::Windows::Forms::ComboBox^ comboBox1;
+
+
+
+
+
+
+
 
 
 
@@ -149,6 +160,7 @@ namespace AgroRobotView {
 			this->textBox1 = (gcnew System::Windows::Forms::TextBox());
 			this->label1 = (gcnew System::Windows::Forms::Label());
 			this->dataGridView1 = (gcnew System::Windows::Forms::DataGridView());
+			this->comboBox1 = (gcnew System::Windows::Forms::ComboBox());
 			this->Column1 = (gcnew System::Windows::Forms::DataGridViewTextBoxColumn());
 			this->Column2 = (gcnew System::Windows::Forms::DataGridViewTextBoxColumn());
 			this->Column3 = (gcnew System::Windows::Forms::DataGridViewTextBoxColumn());
@@ -162,10 +174,9 @@ namespace AgroRobotView {
 			// 
 			// button4
 			// 
-			this->button4->Location = System::Drawing::Point(585, 439);
-			this->button4->Margin = System::Windows::Forms::Padding(4, 4, 4, 4);
+			this->button4->Location = System::Drawing::Point(384, 357);
 			this->button4->Name = L"button4";
-			this->button4->Size = System::Drawing::Size(100, 28);
+			this->button4->Size = System::Drawing::Size(75, 23);
 			this->button4->TabIndex = 27;
 			this->button4->Text = L"Eliminar";
 			this->button4->UseVisualStyleBackColor = true;
@@ -173,10 +184,9 @@ namespace AgroRobotView {
 			// 
 			// button3
 			// 
-			this->button3->Location = System::Drawing::Point(389, 439);
-			this->button3->Margin = System::Windows::Forms::Padding(4, 4, 4, 4);
+			this->button3->Location = System::Drawing::Point(174, 357);
 			this->button3->Name = L"button3";
-			this->button3->Size = System::Drawing::Size(100, 28);
+			this->button3->Size = System::Drawing::Size(75, 23);
 			this->button3->TabIndex = 26;
 			this->button3->Text = L"Editar";
 			this->button3->UseVisualStyleBackColor = true;
@@ -184,10 +194,9 @@ namespace AgroRobotView {
 			// 
 			// button2
 			// 
-			this->button2->Location = System::Drawing::Point(180, 439);
-			this->button2->Margin = System::Windows::Forms::Padding(4, 4, 4, 4);
+			this->button2->Location = System::Drawing::Point(535, 57);
 			this->button2->Name = L"button2";
-			this->button2->Size = System::Drawing::Size(100, 28);
+			this->button2->Size = System::Drawing::Size(75, 23);
 			this->button2->TabIndex = 25;
 			this->button2->Text = L"Agregar";
 			this->button2->UseVisualStyleBackColor = true;
@@ -198,21 +207,18 @@ namespace AgroRobotView {
 			this->groupBox1->Controls->Add(this->button1);
 			this->groupBox1->Controls->Add(this->textBox1);
 			this->groupBox1->Controls->Add(this->label1);
-			this->groupBox1->Location = System::Drawing::Point(239, 31);
-			this->groupBox1->Margin = System::Windows::Forms::Padding(4, 4, 4, 4);
+			this->groupBox1->Location = System::Drawing::Point(49, 24);
 			this->groupBox1->Name = L"groupBox1";
-			this->groupBox1->Padding = System::Windows::Forms::Padding(4, 4, 4, 4);
-			this->groupBox1->Size = System::Drawing::Size(424, 123);
+			this->groupBox1->Size = System::Drawing::Size(318, 100);
 			this->groupBox1->TabIndex = 23;
 			this->groupBox1->TabStop = false;
 			this->groupBox1->Text = L"Criterio de Busqueda";
 			// 
 			// button1
 			// 
-			this->button1->Location = System::Drawing::Point(280, 43);
-			this->button1->Margin = System::Windows::Forms::Padding(4, 4, 4, 4);
+			this->button1->Location = System::Drawing::Point(210, 35);
 			this->button1->Name = L"button1";
-			this->button1->Size = System::Drawing::Size(100, 28);
+			this->button1->Size = System::Drawing::Size(75, 23);
 			this->button1->TabIndex = 4;
 			this->button1->Text = L"Buscar";
 			this->button1->UseVisualStyleBackColor = true;
@@ -220,19 +226,17 @@ namespace AgroRobotView {
 			// 
 			// textBox1
 			// 
-			this->textBox1->Location = System::Drawing::Point(89, 43);
-			this->textBox1->Margin = System::Windows::Forms::Padding(4, 4, 4, 4);
+			this->textBox1->Location = System::Drawing::Point(67, 35);
 			this->textBox1->Name = L"textBox1";
-			this->textBox1->Size = System::Drawing::Size(132, 22);
+			this->textBox1->Size = System::Drawing::Size(100, 20);
 			this->textBox1->TabIndex = 1;
 			// 
 			// label1
 			// 
 			this->label1->AutoSize = true;
-			this->label1->Location = System::Drawing::Point(8, 47);
-			this->label1->Margin = System::Windows::Forms::Padding(4, 0, 4, 0);
+			this->label1->Location = System::Drawing::Point(6, 38);
 			this->label1->Name = L"label1";
-			this->label1->Size = System::Drawing::Size(67, 16);
+			this->label1->Size = System::Drawing::Size(55, 13);
 			this->label1->TabIndex = 0;
 			this->label1->Text = L"ID Animal:";
 			// 
@@ -243,12 +247,19 @@ namespace AgroRobotView {
 				this->Column1,
 					this->Column2, this->Column3, this->Column4, this->Column5, this->Column6, this->Column7
 			});
-			this->dataGridView1->Location = System::Drawing::Point(45, 188);
-			this->dataGridView1->Margin = System::Windows::Forms::Padding(4, 4, 4, 4);
+			this->dataGridView1->Location = System::Drawing::Point(34, 153);
 			this->dataGridView1->Name = L"dataGridView1";
 			this->dataGridView1->RowHeadersWidth = 51;
-			this->dataGridView1->Size = System::Drawing::Size(755, 222);
+			this->dataGridView1->Size = System::Drawing::Size(566, 180);
 			this->dataGridView1->TabIndex = 28;
+			// 
+			// comboBox1
+			// 
+			this->comboBox1->FormattingEnabled = true;
+			this->comboBox1->Location = System::Drawing::Point(393, 58);
+			this->comboBox1->Name = L"comboBox1";
+			this->comboBox1->Size = System::Drawing::Size(121, 21);
+			this->comboBox1->TabIndex = 30;
 			// 
 			// Column1
 			// 
@@ -259,7 +270,7 @@ namespace AgroRobotView {
 			// 
 			// Column2
 			// 
-			this->Column2->HeaderText = L"ID Animal";
+			this->Column2->HeaderText = L"Especie";
 			this->Column2->MinimumWidth = 6;
 			this->Column2->Name = L"Column2";
 			this->Column2->Width = 125;
@@ -301,15 +312,15 @@ namespace AgroRobotView {
 			// 
 			// frmMantSangre
 			// 
-			this->AutoScaleDimensions = System::Drawing::SizeF(8, 16);
+			this->AutoScaleDimensions = System::Drawing::SizeF(6, 13);
 			this->AutoScaleMode = System::Windows::Forms::AutoScaleMode::Font;
-			this->ClientSize = System::Drawing::Size(853, 498);
+			this->ClientSize = System::Drawing::Size(640, 405);
+			this->Controls->Add(this->comboBox1);
 			this->Controls->Add(this->dataGridView1);
 			this->Controls->Add(this->button4);
 			this->Controls->Add(this->button3);
 			this->Controls->Add(this->button2);
 			this->Controls->Add(this->groupBox1);
-			this->Margin = System::Windows::Forms::Padding(4, 4, 4, 4);
 			this->Name = L"frmMantSangre";
 			this->Text = L"AgroRobot - Gestion de Muestras de Sangre";
 			this->groupBox1->ResumeLayout(false);
@@ -319,19 +330,60 @@ namespace AgroRobotView {
 
 		}
 #pragma endregion
+	private: void cargarAnimalesDisponibles() {
+		this->comboBox1->Items->Clear();
+
+		// Obtener todos los animales
+		List<Animal^>^ todosAnimales = gestorNutricionalController->obtenerTodosAnimales();
+		// Obtener todas las muestras
+		List<Muestra^>^ todasMuestras = muestraController->buscarTodasMuestrasArchivo(gestorNutricionalController);
+
+		for each (Animal ^ a in todosAnimales) {
+			bool tieneSangre = false;
+			for each (Muestra ^ m in todasMuestras) {
+				if (m->getAnimal() != nullptr &&
+					m->getAnimal()->IdAnimal == a->IdAnimal &&
+					m->getTipo()->Equals("Sangre")) {
+					tieneSangre = true;
+					break;
+				}
+			}
+			if (!tieneSangre) {
+				this->comboBox1->Items->Add(a->IdAnimal);
+			}
+		}
+	}
+
 	private: System::Void button2_Click(System::Object^ sender, System::EventArgs^ e) {
-		frmNuevoMuestraS^ nuevaMuestraSangre = gcnew frmNuevoMuestraS();
+		if (this->comboBox1->SelectedItem == nullptr) {
+			MessageBox::Show("Seleccione un animal válido antes de agregar una muestra de sangre.");
+			return;
+		}
+
+		int idAnimalSeleccionado = Convert::ToInt32(this->comboBox1->SelectedItem->ToString());
+
+		// Abrir el formulario de nueva muestra de sangre
+		frmNuevoMuestraS^ nuevaMuestraSangre = gcnew frmNuevoMuestraS(
+			idAnimalSeleccionado,
+			this->muestraController,
+			this->gestorNutricionalController
+		);
+
 		nuevaMuestraSangre->ShowDialog();
+
+		// Refrescar comboBox después de agregar
+		cargarAnimalesDisponibles();
 	}
 	private: System::Void button1_Click(System::Object^ sender, System::EventArgs^ e) {
+		String^ textoId = this->textBox1->Text->Trim();
+		int idAnimal;
+
 		//MuestraController^ controller = gcnew MuestraController();
 		List<Muestra^>^ listaFiltrada = gcnew List<Muestra^>();
 
-		String^ textoId = this->textBox1->Text->Trim();
-
 		if (String::IsNullOrEmpty(textoId)) {
 			// Mostrar todas las muestras de sangre
-			List<Muestra^>^ todas = muestraController->buscarTodasMuestrasArchivo();
+			List<Muestra^>^ todas = muestraController->buscarTodasMuestrasArchivo(gestorNutricionalController);
 			for each (Muestra ^ m in todas) {
 				if (m->getTipo()->Equals("Sangre")) {
 					listaFiltrada->Add(m);
@@ -339,13 +391,12 @@ namespace AgroRobotView {
 			}
 		}
 		else {
-			// Buscar por ID de animal
-			int idAnimal;
+
 			if (!Int32::TryParse(textoId, idAnimal)) {
 				MessageBox::Show("Ingrese un ID de animal válido.");
 				return;
 			}
-			List<Muestra^>^ listaMuestras = muestraController->buscarMuestrasPorAnimalArchivo(idAnimal);
+			List<Muestra^>^ listaMuestras = muestraController->buscarMuestrasPorAnimalArchivo(idAnimal, gestorNutricionalController);
 			for each (Muestra ^ m in listaMuestras) {
 				if (m->getTipo()->Equals("Sangre")) {
 					listaFiltrada->Add(m);
@@ -367,7 +418,7 @@ namespace AgroRobotView {
 			Muestra^ m = listaMuestras[i];
 			array<String^>^ filaGrilla = gcnew array<String^>(7);
 			filaGrilla[0] = Convert::ToString(m->getIdMuestra());
-			filaGrilla[1] = Convert::ToString(m->getIdAnimal());
+			filaGrilla[1] = m->getEspecie();
 			filaGrilla[2] = m->getCantidadExtraida();
 			filaGrilla[3] = m->getCoagulos();
 			filaGrilla[4] = m->getContaminacion();
@@ -378,15 +429,23 @@ namespace AgroRobotView {
 	}
 	
 	private: System::Void button3_Click(System::Object^ sender, System::EventArgs^ e) {
+		if (this->dataGridView1->SelectedRows->Count == 0) {
+			MessageBox::Show("Seleccione una fila para editar.");
+			return;
+		}
+
 		int filaSeleccionada = this->dataGridView1->SelectedRows[0]->Index;
 		int idMuestraEditar = Convert::ToInt32(this->dataGridView1->Rows[filaSeleccionada]->Cells[0]->Value->ToString());
 
-		Muestra^ muestra = this->muestraController->buscarMuestraPorIdArchivo(idMuestraEditar);
-		if (muestra == nullptr || muestra->getTipo() != "Sangre") {
+		// Usar el gestor para reconstruir la muestra correctamente
+		Muestra^ muestra = this->muestraController->buscarMuestraPorIdArchivo(idMuestraEditar, this->gestorNutricionalController);
+
+		if (muestra == nullptr || !muestra->getTipo()->Equals("Sangre")) {
 			MessageBox::Show("La muestra seleccionada no es válida o no es de tipo sangre.");
 			return;
 		}
 
+		// Abrir formulario de edición de muestra de sangre
 		frmEditarMuestraS^ ventanaEditar = gcnew frmEditarMuestraS(muestra, this->muestraController);
 		ventanaEditar->ShowDialog();
 	}
@@ -399,8 +458,28 @@ namespace AgroRobotView {
 		int filaSeleccionada = this->dataGridView1->SelectedRows[0]->Index;
 		int idEliminar = Convert::ToInt32(this->dataGridView1->Rows[filaSeleccionada]->Cells[0]->Value->ToString());
 
-		this->muestraController->eliminarMuestraArchivo(idEliminar);
-		MessageBox::Show("La muestra ha sido eliminada con éxito.");
+		// Confirmación antes de eliminar
+		System::Windows::Forms::DialogResult resultado =
+			MessageBox::Show("¿Está seguro de que desea eliminar la muestra seleccionada?",
+				"Confirmar eliminación",
+				MessageBoxButtons::YesNo,
+				MessageBoxIcon::Warning);
+
+		if (resultado == System::Windows::Forms::DialogResult::Yes) {
+			this->muestraController->eliminarMuestraArchivo(idEliminar);
+			MessageBox::Show("La muestra ha sido eliminada con éxito.");
+
+			// Refrescar la grilla mostrando solo muestras de sangre
+			List<Muestra^>^ todas = this->muestraController->buscarTodasMuestrasArchivo(this->gestorNutricionalController);
+			List<Muestra^>^ listaFiltrada = gcnew List<Muestra^>();
+			for each (Muestra ^ m in todas) {
+				if (m->getTipo()->Equals("Sangre")) {
+					listaFiltrada->Add(m);
+				}
+			}
+			this->dataGridView1->Rows->Clear();
+			mostrarGrilla(listaFiltrada);
+		}
 	}
 };
 }
